+++ zcash_client_backend/src/data_api/testing.rs
@@ -1071,7 +1071,6 @@ where
             &SpendingKeys::from_unified_spending_key(usk.clone()),
             ovk_policy,
             &proposal,
-            None,
         )
     }
 
@@ -1320,7 +1319,6 @@ where
             &SpendingKeys::from_unified_spending_key(usk.clone()),
             ovk_policy,
             proposal,
-            None,
         )
     }